Consolidate test files for better organization and maintainability
